Add files via upload
Car Park
A classic example of the application of Programmable Logic Controllers (PLCs) is in automated public car parks. While the technology has developed in terms of image recognition of number plates the fundamental principle stays the same. If appropriate entry conditions are met on entry, entry is allowed and if appropriate conditions are met on exit, exit is allowed.
Perform a review of car park control system technologies. This review should include a discussion of alternative control strategies to that described in the design brief. You should then review appropriate controller, sensors, and actuators and their operating principles. You should make an informed selection on the most appropriate type for your design brief. 
Brief
To design the control system for a small car park capable of holding 10 cars. The design brief is as follows:
•	On entry
o	If there are less than 10 cars in the car park only a green light will illuminate, and entry can be permitted. 
o	If there are between 9 and 13 cars (not inclusive) only an orange light will illuminate, and entry can be permitted. 
o	Otherwise, no entry is permitted. 
o	Entry will only be permitted in the event of the detection of the presence of a car and a ticket request, usually via a push button.
•	On exit
o	Exit will be permitted if both the ticket is shown to be valid and the presence of a car is detected.
o	To simplify the problem, you can treat ticket validity as a single bit.